requirements.txt
httpx>=0.27.0            # Async HTTP client for CRM PATCH calls

# Core
supabase==2.28.3

# Auth
PyJWT==2.10.1
cryptography==43.0.3

# Utilities
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
psycopg2-binary==2.9.9

# Web
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9

# Comms
twilio==9.4.3
sendgrid==6.11.0
requests==2.32.3

groq==0.9.0
sentence-transformers==3.0.1
